+++ db/schema.rb
@@ -27,8 +27,8 @@ ActiveRecord::Schema.define(version: 20150307010121) do
     t.integer  "updater_id"
     t.integer  "deleter_id"
     t.datetime "deleted_at"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                 null: false
+    t.datetime "updated_at",                 null: false
   end
 
   create_table "audio_events", force: :cascade do |t|