+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+ - 2015-03-09
+    - More changes to project and site pages [#164]
+
  - 2015-03-08
     - Improved site lat/long obfuscation calculation
     - Refactored permission code to prepare for project logged in and anon permissions (this was a quite large and sweeping change) [#99](https://github.com/QutBioacoustics/baw-server/issues/99)